Fix migration to include knowledge of tag slugs
Because the branch which had the 'slug' field added to the EventTag
model was merged into the branch which had the event pricing/title
fields added, the migrations were manually merged, and the 'slug' field
addition was missed. This meant that if auto schema migration is used
south thinks that it needs to add the slug field, even though it's
already there.